terafamiliar/requirements.txt
asgiref==2.3.2
astroid==2.0.4
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
certifi==2018.8.24
colorama==0.3.9
constantly==15.1.0
daphne==2.2.2
Django==2.1.2
django-ckeditor==5.6.1
django-cleanup==3.0.1
django-js-asset==1.1.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
PyHamcrest==1.9.0
pylint==2.1.1
pytz==2018.5
six==1.11.0
txaio==18.8.1
wincertstore==0.2
wrapt==1.10.11
zope.interface==4.5.0
